visibility as float + ambient light

+++ PER.Abstractions/src/Environment/LevelObject.cs
@@ -67,7 +67,7 @@ public abstract class LevelObject<TLevel, TChunk, TObject>
     }
 
     internal List<ILight?>? blockedLights { get; private set; }
-    internal Dictionary<Vector2Int, (float lighting, int visibility)>? contributedLight { get; private set; }
+    internal Dictionary<Vector2Int, (float lighting, float visibility)>? contributedLight { get; private set; }
 
     private TLevel? _level;
 
@@ -81,7 +81,7 @@ public abstract class LevelObject<TLevel, TChunk, TObject>
             if(blocksLight)
                 blockedLights ??= new List<ILight?>();
             if(this is ILight)
-                contributedLight ??= new Dictionary<Vector2Int, (float, int)>();
+                contributedLight ??= new Dictionary<Vector2Int, (float, float)>();
         }
         else {
             blockedLights?.Clear();
